third_party/WebKit/LayoutTests/fast/loader/scroll-restore-on-back-ordering.html

<body>
<div style="height: 1000px"></div>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

window.onpopstate = function() {
    console.log("popstate " + window.scrollY);
}

window.onhashchange = function() {
    console.log("hashchanged " + window.scrollY);
    if (window.testRunner)
        testRunner.notifyDone();
}

window.onload = function() {
    setTimeout(function() {
        window.scrollTo(0, 1000);
        history.pushState({ }, '', '#a');
        window.scrollTo(0, 0);
        history.back();
    }, 0);
}

</script>
</body>
